+++ packages/blocks/CHANGELOG.md
@@ -3,3 +3,7 @@
 ### Breaking Change
 
 - The `isSharedBlock` function is removed. Use `isReusableBlock` instead.
+
+### Deprecations
+
+- The `getDefaultBlockForPostFormat` function has been deprecated.